dp-service #197: verify bucket spans in every service that queries buckets
The annotation service reaches the same bucket overlap filter as the query
service, through executeDataBlockQuery in the dataset export jobs, but did
not run the startup verification. Because the flag controlling the query
lower bound is a process-wide static, that process applied the bound against
an archive it had never checked, and a legacy over-long bucket would be
silently omitted from exports.

Moves verifyBucketSpans() from MongoSyncQueryClient to the shared
MongoSyncClient and calls it from MongoAnnotationHandler.init_() as well as
MongoQueryHandler.init_(), so any service that issues bucket time-range
queries establishes the invariant before relying on it.
