+++ CLAUDE.md
@@ -266,7 +266,8 @@ The record being updated is excluded from the check via `Filters.ne(clientActiva
 `Buckets.maxBucketSpanSeconds` is a shared invariant between ingestion and query, and both of its failure modes are *silent wrong answers* rather than errors — treat it accordingly:
 - **`BucketSpanLimits`** — single source for the limit. Value is validated once (rejects non-positive, and anything above `MAX_CONFIGURABLE_SPAN_SECONDS` where the nanos conversion would overflow) and cached; invalid config throws `DpRuntimeException`.
 - **Ingestion** enforces the limit for *new* data only, via `IngestionValidationUtility`.
-- **Query** adds the `firstTime` lower bound only when the stored archive is known to comply. `BucketSpanVerifier` checks this at query-service startup and records the outcome in the `bucketSpanVerification` collection, so the scan runs once per limit value rather than every restart. On violation or error the bound is disabled process-wide (`BucketSpanLimits.disableQueryLowerBound()`) and queries degrade to the slower unbounded scan — correct but slow, never fast but wrong.
+- **Query** adds the `firstTime` lower bound only when the stored archive is known to comply. `BucketSpanVerifier` checks this and records the outcome in the `bucketSpanVerification` collection, so the scan runs once per limit value rather than every restart. On violation or error the bound is disabled process-wide (`BucketSpanLimits.disableQueryLowerBound()`) and queries degrade to the slower unbounded scan — correct but slow, never fast but wrong.
+- `verifyBucketSpans()` lives on `MongoSyncClient` because the flag it controls is **process-wide** and more than one service issues bucket time-range queries: the query service directly, and the annotation service through dataset export (`executeDataBlockQuery`). Any new service that queries buckets must call it from its handler's `init_()`, or that process will apply the bound unverified.
 - Disable the check with `Buckets.verifyBucketSpansOnStartup: false` only when compliance has been confirmed independently. Off by default under test.
 - **Never sample** as a shortcut for this check: over-long buckets are typically rare, so a sample that misses them reports a false all-clear.
