+++ src/main/java/com/ospreydcs/dp/service/annotation/handler/mongo/MongoAnnotationHandler.java
@@ -59,6 +59,12 @@ public class MongoAnnotationHandler extends QueueHandlerBase implements Annotati
         if (!mongoQueryClient.init()) {
             logger.error("error in mongoQueryClient.init");
         }
+
+        // Dataset export issues bucket time-range queries through executeDataBlockQuery, so this
+        // process must establish the max bucket span invariant before relying on the query lower
+        // bound (#197) just as the query service does. The flag it controls is process-wide.
+        mongoQueryClient.verifyBucketSpans();
+
         return true;
     }
